requirements.txt
streamlit==1.43.0
# git+https://github.com/marks/streamlit-airtable-connection
# st-annotated-text==4.0.1
streamlit-bridge==1.1.3
pydantic>=2.6.2
# langchain_core
python-dateutil==2.8.2
dateparser
matplotlib==3.8.3
qrcode==7.4.2
vobject==0.9.6.1
Pillow
streamlit-tags==1.2.8
python-docx==1.2.0
reportlab==4.4.4
phonenumbers==9.0.19